Fix delegate over-release bug in OptionsViewController; Fix paths & header include references for clipper
The delegate is assigned (weak reference) by the OptionsViewController.  It should not be released.
I haven't seen this bug manifest itself, but it would likely cause a crash in any case where our view
controller gets cleaned up and we try to set a message to the delegate.

I guess we need a better way to deal with project dependencies.  I was under the impression that clipper lives in 'utilties'.
Should I move it somewhere else?